Поправил местами код и pom.xml  Изменил использование имплементации сервиса на интерфейс  Убрал модификаторы доступа у интерфейса  Сделал id у Role Long, чтобы у всех был только Long (выглядело странно)  Добавил pluginManagement для правильной сборки проекта через докер в плагинах pom.xml Вернул миграцию БД  Межсервисное общение будет с миграциями происходить через схемы БД, чтобы были раздельные истории Flyway и не было пересечений Добавил профили Spring для dev (разные порты, только localhost) и prod (фиксированные порты, внутренние адреса кубера) Добавил в докерфайл возможность самостоятельной сборки проекта

+++ src/main/java/com/perflibnetcracker/authenticationservice/model/Role.java
@@ -1,10 +1,19 @@
 package com.perflibnetcracker.authenticationservice.model;
 
 
-import com.fasterxml.jackson.annotation.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToMany;
+import javax.persistence.Table;
 import java.util.Set;
 
 @Entity
@@ -17,7 +26,7 @@ import java.util.Set;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
